Applied patch to add all hosts to ACL dring VM initialization
I'm also attaching that patch which fixes a problem that we found during our testing. The issue was that the host which is running the VM does not necessarily intialize the volume initially. It is usually the master that initializes the volume and then the host where the VM is running does the iscsi login and attaches the disk to the VM. This fix will basically add all the hosts in a cluster to the ACL.